added comment and improved file checking

+++ the_mines/get_html.py
@@ -1,47 +1,34 @@
-from argparse import ArgumentParser
 import requests
 import shutil
+import os
 
 
-def get_html():
-    parser = ArgumentParser()
-    parser.add_argument(
-        "--matchday", action="store", required=True, help="ex. 8",
-    )
-    parser.add_argument(
-        "--season", action="store", required=True, help="ex. 2019/2020",
-    )
-    args = parser.parse_args()
+def get_html(matchday, season):
 
-    matchday = f"{args.matchday}"
-    season = f"{args.season}"
-
+    # - The URL of this website uses only the latter year in a season (2019/2020).
+    # - season[5:9] selects the latter season.
+    # - I'm using the 2019/2020 convention for season input because I believe it
+    #   is what most users are familiar with when asked for a season
     url = "https://www.fussballdaten.de/bundesliga/" + season[5:9] + "/" + matchday
 
     page_source = requests.get(url)
     season = season.replace("/", "_")
     filename = "fussballdaten_" + season + "_" + matchday + ".html"
+    filepath = 'data/' + filename
 
     # Check if file exists
-    try:
-        f = open("data/" + filename)
+    if not os.path.isfile(filepath):
+        with open(filename, 'w') as f:
+            print("creating file..." + filename)
+            f.write(str(page_source.content))
+            shutil.move(filename, "data/" + filename)
+    else:
         print("opening file... " + filename)
         return filename
-    except IOError:
-        # if not, then create it
-        f = open(filename, "w")
-        print("creating file..." + filename)
-        f.write(str(page_source.content))
-        shutil.move(filename, "data/" + filename)
-    finally:
-        f.close()
 
-    return filename
-
-
-def main():
-    get_html()
+    return filename # should this be in the if statement or out here?
 
 
 if __name__ == "__main__":
-    main()
+
+    get_html(matchday, season)
